Address Claude strict-review pass 5
- fused_mhc_kernels: H_post bias gradient reduction now upcasts to fp32
  before summing across `s*b`, matching the rest of the mHC fp32-accumulation
  practice (was IMPORTANT bf16 reduction error for long sequences).
- hyper_connection.output_contract: assert n-stream input dim is divisible
  by num_residual_streams; fails loud instead of silently producing a wrong
  C via integer division.
- HyperConnectionHybridLayer: pass `training=self.training` to the
  layer-boundary BDA. `dropout_prob=0.0` already disables dropout, but the
  hard-coded `False` was misleading inside a training-mode forward.
- CheckpointManager: introduce a `_finalized` flag. add_checkpoint() and
  discard_all_outputs_and_register_unified_recompute() now raise on reuse
  after finalization, matching the documented single-use contract.
- HyperConnectionTransformerLayer.__call__: comment documents the
  single-thread assumption behind the stash-on-self pattern; concurrent
  inference would race on `_mhc_recompute_manager` and would need a
  thread-local or kwarg-passing redesign.
- TransformerBlock._build_mhc_recompute_layer_plan: deterministic
  block-end markers are now computed once and cached on the instance via
  `_compute_mhc_block_end_plan`; per-forward work is just the
  CheckpointManager allocation, not the O(num_layers) layout loop.

+++ megatron/core/fusions/fused_mhc_kernels.py
@@ -539,7 +539,10 @@ if _CUTILE_AVAILABLE:
                     TILE_SIZE,
                 ),
             )
-        g_bias = g_x.sum(dim=0) if bias is not None else None
+        # Accumulate the bias gradient in fp32 to match the rest of the mHC
+        # implementation; bf16 reduction across `s*b` rows accumulates noticeable
+        # rounding error for long sequences / large batches.
+        g_bias = g_x.float().sum(dim=0).to(g_x.dtype) if bias is not None else None
         return (
             g_hr.view(s, b, n, n),
             g_res.view(s, b, n, C),
